refactor: FakeLottoNumberGenerator를 하나의 파일에서 관리하도록 수정한다.
