--- src/test/kotlin/lotto/strategy/FifthRankLottoNumberGenerator.kt
@@ -1,9 +1,0 @@
-package lotto.strategy
-
-import lotto.stretagy.LottoNumberListGenerator
-
-class FifthRankLottoNumberGenerator : LottoNumberListGenerator {
-    override fun generate(): List<Int> {
-        return listOf(1, 2, 3, 43, 44, 45)
-    }
-}
