--- src/test/kotlin/lotto/strategy/FourRankLottoNumberGenerator.kt
@@ -1,9 +1,0 @@
-package lotto.strategy
-
-import lotto.stretagy.LottoNumberListGenerator
-
-class FourRankLottoNumberGenerator : LottoNumberListGenerator {
-    override fun generate(): List<Int> {
-        return listOf(1, 2, 3, 4, 44, 45)
-    }
-}
