--- src/test/kotlin/lotto/strategy/ThirdRankLottoNumberGenerator.kt
@@ -1,9 +1,0 @@
-package lotto.strategy
-
-import lotto.stretagy.LottoNumberListGenerator
-
-class ThirdRankLottoNumberGenerator : LottoNumberListGenerator {
-    override fun generate(): List<Int> {
-        return listOf(1, 2, 3, 4, 5, 7)
-    }
-}
